refactor(languages): Onboard Japanese to new module layout
