+++ classes/jbook.lua
@@ -8,9 +8,10 @@ class._name = "jbook"
 
 function class:_init (options)
    tbook._init(self, options)
-   SILE.languageSupport.loadLanguage("ja")
-   SILE.settings:set("document.language", "ja", true)
    SILE.settings:set("font.family", "Noto Sans CJK JP", true)
+   self:registerPostinit(function (_)
+      SILE.settings:set("document.language", "ja", true)
+   end)
 end
 
 class.declareOptions = tbook.declareOptions
