+++ languages/ja/hyphens.lua
@@ -1,0 +1,1 @@
+return {}
